Changes to GoalsDashboard / by jameshobson

+++ src/main/java/co/grandcircus/dreamsetter/dao/GoalDao.java
@@ -54,7 +54,7 @@ public class GoalDao {
 		}
 	}
 
-	public int addGoal(Goal goal) {
+	public double addGoal(Goal goal) {
 		String sql = "INSERT INTO AddGoal (goalAmount, goalDescription , startDate, dueDate) VALUES (?, ?, ? ,?)";
 		try (Connection connection = (connectionFactory).getConnection();
 				PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
